# Cache RawData for Commit, Tag, & Tree, fixes #6