A `rawdata []byte` member is added to each type. It is computed on the first call to `RawData()`.
Also noticed a few odd indentations and fixed them.

Let me know if there are any changes that need to be made, feedback is welcome.