@sameer 2018-12-18
Sure, should that be done as a `*testing.B` in `git_test.go`, or should I just try it out against the linux kernel or something with a before & after?