@magik6k 2018-12-18
Whatever works best for you and you think should show a difference is fine.