@sameer 2018-12-18
Ok, I had a chance to do more comprehensive testing. All the files in the repo are already added to my ipfs daemon, so they should return fast. The bench script just does `rm -rf .git/ipld/ && git push ipld:: master`.
Before: 
```
 Performance counter stats for './bench.sh' (100 runs):

            554.14 msec task-clock:u              #    0.623 CPUs utilized            ( +-  0.85% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             2,680      page-faults:u             # 4840.519 M/sec                    ( +-  1.63% )
     1,121,182,658      cycles:u                  # 2025221.108 GHz                   ( +-  0.29% )
     1,280,351,427      instructions:u            #    1.14  insn per cycle           ( +-  0.13% )
       227,040,471      branches:u                # 410109049.439 M/sec               ( +-  0.14% )
         4,581,354      branch-misses:u           #    2.02% of all branches          ( +-  0.32% )

            0.8896 +- 0.0205 seconds time elapsed  ( +-  2.31% )
```
After:
```
 Performance counter stats for './bench-fast.sh' (100 runs):

            567.96 msec task-clock:u              #    0.646 CPUs utilized            ( +-  0.86% )
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
             2,678      page-faults:u             # 4719.887 M/sec                    ( +-  1.80% )
     1,131,191,140      cycles:u                  # 1993463.988 GHz                   ( +-  0.33% )
     1,281,825,576      instructions:u            #    1.13  insn per cycle           ( +-  0.14% )
       227,317,086      branches:u                # 400594036.690 M/sec               ( +-  0.16% )
         4,627,311      branch-misses:u           #    2.04% of all branches          ( +-  0.31% )

            0.8797 +- 0.0203 seconds time elapsed  ( +-  2.31% )
```

There doesn't really seem to be any difference when I did this test. Is there some better way to do one?