@magik6k 2018-12-19
Most of this will be I/O. I'd use `testing.B`, load test data like in https://github.com/ipfs/go-ipld-git/blob/master/git_test.go#L22, and then call RawData in a loop `b.N` times